Fallback to local (ollama) model when Azure credentials not specified
In case the .env file didn't contain an Azure API key and resource name you couldn't test the chatbot's and the session would error out.

I added a `getDefaultModel()` which replaced all instances of using constant `'azure/gpt-5-mini'` as a defualt model, this way you can control which model to use by default, I made it use ollama and Llama3.2 by default but this could be changed by playing with the `getDefaultModel()` function and adding the appropriate provider for your model (an ollama one already exitsted in the codebase but wasn't used)

You can also select Llama 3.2 (Ollama) as a base model from the selection menu for models, this is also to prevent it from crashing by not using Azure hosted models.